### 2025-06-18 #8423 @editorialbot
[PRE REVIEW]: ForeFire: A Modular, Scriptable C++ Simulation Engine and Library for Wildland-Fire Spread 
**Submitting author:** @filippi (<a href="http://orcid.org/0000-0002-6244-0648">Jean-Baptiste Filippi</a>)
**Repository:** https://github.com/forefireAPI/forefire/
**Branch with paper.md** (empty if default branch):
**Version:** v2.0
