### 2025-06-24 bf90ae9
Paper made shorter
