+++ paper/paper.md
@@ -85,72 +85,47 @@ bibliography: paper.bib
 
 # Summary
 Wildfire forecasting is at the same time an active research area and an important need for operational decision support systems. **ForeFire** is a modular, high-performance wildland fire simulation engine implemented in C++. It is designed to model the spread of wildfire perimeters over large landscapes at meter scale resolution, serving both as a research platform and an operational forecasting tool. 
-ForeFire can be used to forecast wildfires spanning thousands of hectares within seconds, supporting wildfire management operations, while also providing an open testbed for experimenting with new fire behavior, fire intensity, and fire flux models in a scientific context.
-
-ForeFire offers multiple interfaces and utilities to maximize usability and integration. 
-The core C++ library has Fortran and Python bindings and is accompanied by a lightweight scriptable interpreter (in custom FF language), a local HTTP service that enables a customizable graphical user interface and ability to load, save and export data in NetCDF, GeoJson, KML, png and jpg. 
-ForeFire is discrete‑event‑driven [@filippi2009], focusing computational effort on the active region of the fire front defined as a dynamic mesh or multipolygons of fire markers. 
-
-Moreover, additionaly ForeFire is also capable of two-way coupling with the MesoNH [@lac2018] atmospheric model to account for fire-atmosphere interaction [@filippi2013].
+It also offers multiple interfaces, the core C++ library has Fortran and Python bindings and is accompanied by a lightweight scriptable interpreter (in custom FF language), a local HTTP service that enables a customizable graphical user interface and ability to load, save and export data in NetCDF, GeoJson, KML, png and jpg. Simulation is discrete‑event‑driven [@filippi2009], focusing computational effort on the active region of the fire front defined as a dynamic mesh or multipolygons of fire markers. ForeFire is also capable of two-way coupling with the MesoNH [@lac2018] atmospheric model to account for fire-atmosphere interaction [@filippi2013].
 
 # Statement of need
 
 Wildfire modeling tools have historically been split between **complex combustion research models** and **streamlined operational tools**, each with distinct limitations. 
 On one end of the spectrum, computational combustion and fluid dynamics (CFD) based models (e.g., FIRETEC [@linn2005] or WFDS [@mell2007]) provide detailed physics with complete control of combustion models at flame scale but are highly computationally intensive and yet unable to provide faster than real time large wildfire forecasting. 
-On the other end, operational wildfire simulators built for practitioners, such as widely used Farsite (now flammap) [@finney1998] or Canadian Prometheus [@garcia2008], while not resolving the combustion at flame scale, are able to simulate tens of kilometers long fire front in matter of seconds, prioritizing speed and simplicity at the expense of generality. These codes have definite built-in modeling assumptions and are distributed (freely) as compiled software with graphical interfaces and limited access to source code. 
-While being well-packaged is advantageous for direct use, it makes it difficult for researchers to implement new modeling approaches or to streamline and script integration into different experimental frameworks. This gap between highly complex customizable models and more rigid operational tools creates a need for an intermediate solution: a wildfire simulator that is both **adaptable** and **high-performing**. 
-ForeFire was developed to fill this need by bridging the flexibility of research-oriented code with the efficiency and user-friendliness of operational systems [@filippi2018].
-
-Other open source libraries for wildland fire spread do exist as for example ElmFire [@lautenberger2013] or Cell2Fire [@pais2021], but with respect to ForeFire, these packages use a different numerical scheme, and are tied to a single spread models and do not provide a scriptable scenario interface or atmospheric coupling. 
[…]
