### 2025-06-24 #8423 @filippi
Le mer. 25 juin 2025, 06:28, Kristen Thyng ***@***.***> a
écrit :

> *kthyng* left a comment (openjournals/joss-reviews#8423)
> <https://github.com/openjournals/joss-reviews/issues/8423#issuecomment-3001942675>
>
> @filippi <https://github.com/filippi> Please reduce to length of your
> paper to up to about 1000 words. Paper format information here:
> https://joss.readthedocs.io/en/latest/paper.html. You can check paper
> length (and a few other items) with the command @editorialbot check
> repository though note it needs to be the first item in a comment to run
> correctly.
>
