+++ paper/paper.md
@@ -84,34 +84,28 @@ bibliography: paper.bib
 ---
 
 # Summary
-Wildfire forecasting is at the same time an active research area and an important need for operational decision support systems. **ForeFire** is a modular, high-performance wildland fire simulation engine implemented in C++. It is designed to model the spread of wildfire perimeters over large landscapes at meter scale resolution, serving both as a research platform and an operational forecasting tool. 
-It also offers multiple interfaces, the core C++ library has Fortran and Python bindings and is accompanied by a lightweight scriptable interpreter (in custom FF language), a local HTTP service that enables a customizable graphical user interface and ability to load, save and export data in NetCDF, GeoJson, KML, png and jpg. Simulation is discrete‑event‑driven [@filippi2009], focusing computational effort on the active region of the fire front defined as a dynamic mesh or multipolygons of fire markers. ForeFire is also capable of two-way coupling with the MesoNH [@lac2018] atmospheric model to account for fire-atmosphere interaction [@filippi2013].
+Wildfire forecasting is at the same time an active research area and an important need for decision support systems. **ForeFire** is a modular, high-performance, scriptable, discrete‑event‑driven simulation engine [@filippi2009]. It is designed to model the spread of wildfire perimeters over large landscapes at meter scale resolution in seconds, serving both as a research platform and an operational forecasting tool. Core C++ library has Fortran and Python bindings and is accompanied by a lightweight scriptable interpreter (in custom FF language), a local HTTP service with customizable graphical user interface and can load, save and export data in NetCDF, GeoJson, KML, png and jpg. ForeFire can also account for fire-atmosphere interaction by two-way coupling with the MesoNH [@lac2018] atmospheric model [@filippi2013].
 
 # Statement of need
 
 Wildfire modeling tools have historically been split between **complex combustion research models** and **streamlined operational tools**, each with distinct limitations. 
-On one end of the spectrum, computational combustion and fluid dynamics (CFD) based models (e.g., FIRETEC [@linn2005] or WFDS [@mell2007]) provide detailed physics with complete control of combustion models at flame scale but are highly computationally intensive and yet unable to provide faster than real time large wildfire forecasting. 
-Another family of wildland fire spread solvers such as WRF/SFire [@mandel2011] can resolve coupled front propagation and local meteorology but must be run within an atmospheric model that requires a large amount of processing power and data.
-
-On the other end, operational wildfire simulators built for practitioners, such as widely used Farsite (now flammap) [@finney1998] or Canadian Prometheus [@garcia2008] are able to simulate tens of kilometers long fire front in matter of seconds, prioritizing speed and simplicity at the expense of generality. These codes have definite built-in modeling assumptions and are distributed (freely) as compiled software with graphical interfaces and limited access to source code. Existing open‑source libraries, such as ElmFire [@lautenberger2013] or Cell2Fire [@pais2021], are each built around a single core spread model and are packaged with streamlined, simplified parameterization to prioritize practitioner efficiency. While being well-packaged is advantageous for direct use, it makes it difficult for researchers to implement new modeling approaches or to streamline and script integration into different experimental frameworks. 
-This gap between highly complex customizable models and more rigid operational tools creates a need for a **unified solution** : a wildfire simulator that is both **adaptable** and **high-performing**. ForeFire was developed to fill this need by bridging the flexibility of research-oriented code with the efficiency and user-friendliness of operational systems [@filippi2018].
[…]
