+++ paper/paper.md
@@ -91,35 +91,31 @@ Wildfire forecasting is at the same time an active research area and an importan
 Wildfire modeling tools have historically been split between **complex combustion research models** and **streamlined operational tools**, each with distinct limitations. 
 Computational combustion and fluid dynamics (CFD) based models (e.g., FIRETEC [@linn2005] or WFDS [@mell2007]) are highly computationally intensive and yet unable to provide faster than real time large wildfire forecasting. 
 Atmospheric coupled codes, such as WRF/SFire [@mandel2011] must be run within an atmospheric and requires a large amount of processing power and data.
-Operational wildfire simulators such as widely used Farsite (now Flammap) [@finney1998] or Canadian Prometheus [@garcia2008] are able to simulate tens of kilometers long fire front in matter of seconds, but have definite built-in modeling assumptions and are distributed (freely) as compiled software with graphical interfaces. While these are well-packaged, limited access to source code makes it difficult for researchers to implement new modeling approaches or to streamline integration into different experimental frameworks. 
-ForeFire was developed to fill this gap between highly complex customizable models and more rigid operational tools : a **unified** wildfire simulator that is both **adaptable** and **high-performing**. 
+Operational wildfire simulators such as widely used Farsite (now Flammap) [@finney1998] or Canadian Prometheus [@garcia2008] are able to simulate tens of kilometers long fire front in matter of seconds, but have definite built-in modeling assumptions and are distributed (freely) as compiled software with graphical interfaces anf hardly scriptable. 
+ForeFire was developed as a community tool to fill the gap between highly complex customizable models and more rigid operational tools : a **unified** wildfire simulator that is both **adaptable** and **high-performing**. 
 
-# Some typical Use Cases
+# Typical Use Cases
 
 ## Rapid prototyping of new spread or flux models
-Code implements several standard fire fluxes and spread rate models, including different variants of Rothermel model [@andrews2018] and Balbi semi‑physical formulation [@balbi2009], and makes it trivial to switch, extend or add to this withbase with a single `.cpp` using any existing model file as template.
+ForeFire implements several standard fire fluxes and spread rate models, such as Rothermel [@andrews2018] or Balbi [@balbi2009], but also makes it trivial to switch, extend or add to this withbase with a single `.cpp` using any existing model file as template.
 Internally data is handled as *layers* that can come from Python NumPy array, supplied from input NetCDF files or generated on the fly by ForeFire (e.g. slope derived from the elevation layer). 
-Developping a Rate Of Spread wildfire model was the original purpose of this simulation code and helped to iterate versions the Balbi Rate Of Spread formulation on real-cases studies in [@balbi2009] and [@santoni2011]. It also served to implement various heat and chemical species flux models used for volcanic eruption in [@filippi2021], plume chemistry [@strada2012] or industrial fires in [@baggio2022]. In addition, the code includes a generic `ANNPropagationModel`, which implements a feedforward artificial neural network (ANN) that expects a pre-trained graph file.
[…]
