+++ paper/paper.md
@@ -84,36 +84,37 @@ bibliography: paper.bib
 ---
 
 # Summary
-Wildfire forecasting is at the same time an active research area and an important need for decision support systems. **ForeFire** is a modular, high-performance, scriptable, discrete‑event‑driven simulation engine [@filippi2009]. It is designed to model the spread of wildfire perimeters over large landscapes at meter scale resolution in seconds, serving both as a research platform and an operational forecasting tool. Core C++ library has Fortran and Python bindings and is accompanied by a lightweight scriptable interpreter (in custom FF language), a local HTTP service with customizable graphical user interface and can load, save and export data in NetCDF, GeoJson, KML, png and jpg. ForeFire can also account for fire-atmosphere interaction by two-way coupling with the MesoNH [@lac2018] atmospheric model [@filippi2013].
+Wildfire forecasting is both an active research area and an important need for decision support systems. **ForeFire** is a modular, high-performance, scriptable, discrete‑event‑driven simulation engine [@filippi2009] focusing computational effort on the active region of a fire front defined as a dynamic mesh (or multipolygons) of fire markers. It is designed to model the spread of wildfire perimeters over large landscapes at meter scale resolution in seconds, serving both as a research platform and a tool for operational forecasting. The core C++ library has Fortran and Python bindings and is accompanied by a lightweight scriptable interpreter (a custom FF language), can load, save and export data in NetCDF, GeoJson, KML, PNG and JPG, and includes a local HTTP service with customizable graphical user interface. ForeFire can also account for fire-atmosphere interaction by two-way coupling with the MesoNH [@lac2018] atmospheric model [@filippi2013].
 
 # Statement of need
 
 Wildfire modeling tools have historically been split between **complex combustion research models** and **streamlined operational tools**, each with distinct limitations. 
-Computational combustion and fluid dynamics (CFD) based models (e.g., FIRETEC [@linn2005] or WFDS [@mell2007]) are highly computationally intensive and yet unable to provide faster than real time large wildfire forecasting. 
-Atmospheric coupled codes, such as WRF/SFire [@mandel2011] must be run within an atmospheric and requires a large amount of processing power and data.
-Operational wildfire simulators such as widely used Farsite (now Flammap) [@finney1998] or Canadian Prometheus [@garcia2008] are able to simulate tens of kilometers long fire front in matter of seconds, but have definite built-in modeling assumptions and are distributed (freely) as compiled software with graphical interfaces anf hardly scriptable. 
-ForeFire was developed as a community tool to fill the gap between highly complex customizable models and more rigid operational tools : a **unified** wildfire simulator that is both **adaptable** and **high-performing**. 
+Computational combustion and fluid dynamics (CFD) based models (e.g., FIRETEC [@linn2005] or WFDS [@mell2007]) are highly computationally intensive and yet unable to provide large wildfire forecasts faster than real time. 
+Atmospheric coupled codes, such as WRF/SFire [@mandel2011] must be run within an atmospheric model and require a large amount of processing power and data.
+Operational wildfire simulators, such as widely used Farsite (now Flammap) [@finney1998], or Canadian Prometheus [@garcia2008], are able to simulate fire fronts spanning tens of kilometers in a matter of seconds, but have definite built-in modeling assumptions and are distributed as compiled software with graphical interfaces with limited scriptability.
+
[…]
