### 2025-07-02 #8423 @filippi
@kthyng thanks a lot, I have reduce paper to 1110 words, It is as concise as it could be. Is it OK ?
