### 2025-07-30 #8423 @danielskatz
> > > I asked D. Katz to be a reviewer a while back but never got confirmation.
> >
> >
> > Hi - I'm confused by this. Do you mean that you asked me to be a reviewer, or you asked me if you could be a reviewer? In either case, I don't think I've been involved in this submission up to this point, as I don't see myself in any of the earlier comments.
>
> I am also confused.
>
> I think you asked on the Workflows.Community slack channel (about 2 years ago) if anyone wanted to be a reviewer for JOSS. I then DM you if the spot was still available but I think you said you already had other people message you to review for JOSS.

Ok, I see. I think I asked at that time about reviewers for a submission I was editing, but then had found enough reviewers before you responded.

I am not associated with this submission, which is being edited by @philipcardiff, so we would have to ask him why he invited you to review it, but it seems likely that it's because the author suggested you as a possible reviewer in https://github.com/openjournals/joss-reviews/issues/8423#issuecomment-2987202749
