### 2025-07-30 #8423 @iperezx
Maybe my initial question should of been on how the list was generated:https://github.com/openjournals/joss-reviews/issues/8423#issuecomment-2987202749
