### 2025-07-30 #8423 @filippi
Hi @iperezx , that is my fault, I saw in some other submissions that a list of potential reviewers may be added in the submission to help in the editing process.
You ended up there just because you were only 4 I used the keyword "fire" in JOSS reviewer system :
https://reviewers.joss.theoj.org/reviewers/search?language=&area_id=&keywords=fire&name=&commit=Update+search
All 4 were on the list.
Thanks again a lot @philipcardiff for accepting the editing job,  @danielskatz for answering and @kthyng for the pre-review, and to all the potential reviewers to may be part of the evaluation process.
Best regards
