### 2025-07-30 #8423 @iperezx
> Hi [@iperezx](https://github.com/iperezx) , that is my fault, I saw in some other submissions that a list of potential reviewers may be added in the submission to help in the editing process. You ended up there just because we were only 4 I used the keyword "fire" in JOSS reviewer system : https://reviewers.joss.theoj.org/reviewers/search?language=&area_id=&keywords=fire&name=&commit=Update+search I picked up the 2 that has some C on the list. Thanks again a lot [@philipcardiff](https://github.com/philipcardiff) for accepting the editing job, [@danielskatz](https://github.com/danielskatz) for answering and [@kthyng](https://github.com/kthyng) for the pre-review, and to all the potential reviewers to may be part of the evaluation process. Best regards

Thanks, that helps.
Sorry for the long thread but yes I will be interested in viewing a reviewer.
