### 2025-08-13 forefireAPI/forefire#100 @xiazeyu
Concerns for Authorship and Contributions
To ensure all contributors are credited in accordance with JOSS's authorship guidelines, I have a couple of questions based on my review of the repository's commit history:

- Primary Contributor(s): I noticed that a significant portion of the commits are attributed to the GitHub account 'forefire'. Could you please clarify if this is an organizational or shared account? If so, it would be helpful to identify the individual developers behind these contributions to ensure they are properly credited for their substantial work.

- Additional Contributions: I also observed some early commits from a user named ludobegins (Ludovic Beghin), though their contributions appear to have been removed later in the project's history. Since this individual is not listed in the author list or in the acknowledgements, could you please provide some context on his role? This will help clarify whether an authorship credit or formal acknowledgement is warranted.

Parent issue: https://github.com/openjournals/joss-reviews/issues/8680
