### 2025-08-13 forefireAPI/forefire#100 @filippi
Hi @xiazeyu  thanks for reviewing and your question,
I (@filippi)  am the dev behind the forefire account, used for most of the pushes except  from the [2014 initial commit](https://github.com/forefireAPI/forefire/commit/5756e869f0a337dd29dab3dd09c6ec58bd710e5c)(an  import from SVN), to check, it is the "organisation" account of the forefireAPI organisation, for which filippi is the owner. I changed local repos quite often (do not really check how I set it up - which pgp key is used), and I have to say that early or even now mastership of git and GitHub is/was pretty low, I actually realized when I submitted the JOSS paper, originally as @forefire too.

 It is not possible to clean-up the commits history, re-affecting names and accounts, so I understand it is quite confusing.
Some close contributors (Federic Bosseur) contributed many by testing and running the code or preferred to send me their codes by email (Albert Alonso-Pinar)  as is my PhD student, and did not have good git proficiency last year.

if you look at the list/number of commits count (generated by the CI-Bot here [in the paper issue](https://github.com/openjournals/joss-reviews/issues/8680#issuecomment-3140227886)  you may have a better overview of the actual names, as they are based on the "git name".. but here too, a bit confusing, here is some help :

   257	antonio-leblanc              == Main dev in Brazil/umgroumeiu, rather recent, most of the documentation, readme's and Bots,  @ludobegins works with him and they shared the same local repo, so he did push under @ludobegins  for some time,
   148	FILIPPI Jean-Baptiste       == @filippi > main dev/owner
    84	CI BOT
    48	Jean-Baptiste Filippi      == FILIPPI Jean-Baptiste
    31	ludobegins                      == works with Antonio
    26	Batti Filippi                      == FILIPPI Jean-Baptiste
    11	filippi_j                                    == FILIPPI Jean-Baptiste
     6	Frederic Allaire
     6	Thoreau Romain
     5	filippi                                 == FILIPPI Jean-Baptiste
     5	ronan                                 == Ronan Paugam
     5	th96r4ss
     4	Ronan                                == Ronan Paugam
     3	Ludovic Beghin               ==  @ludobegins
     3	baggiocolpir                   ==  Roberta Baggio
     3	jdurand
     2	Antonio Leblanc            == antonio-leblanc
     2	Damien Grandi
     2	Dominik Laux
     2	ronan paugam                   == Ronan Paugam
     1	BAGGIO_R                         == baggiocolpir
     1	Jean-Baptiste FILIPPI.    == FILIPPI Jean-Baptiste
     1	Johanna Strebl
     1	damien
     1	dependabot[bot]
     1	rbaggio                             == Roberta Baggio
     1	serbal193

thanks a lot
Jean-Baptiste
