### 2025-08-14 forefireAPI/forefire#100 @xiazeyu
Hello filippi

Thanks for explaining this. I understand it’s a historic problem that makes the contribution confused. Having a good commit practice will be helpful sometimes when people need to track back who wrote the code to decide who to make contact with. This will be especially useful when your project is growing and more contributors come in to help.

Saying that, your answer solved my concern. I think there's no issue in authorship and contribution.

Thank you
