### 2025-08-24 forefireAPI/forefire#101 @xiazeyu
Wiki is broken
The [wiki](https://forefire.univ-corse.fr/wiki/doku.php) linked from the [project website](https://forefire.univ-corse.fr/) is broken.

Could you please have a check?

Ref https://github.com/openjournals/joss-reviews/issues/8680
