### 2025-08-24 forefireAPI/forefire#101 @xiazeyu
@filippi Thanks for the clarification! The [website](forefire.univ-corse.fr) and [online demo](forefire.univ-corse.fr/sim) is awesome! But it looks like the website itself is significantly outdated, causing trouble when people want to find information about the project.

I suggest to improve the website by:

- Fix the [document link](https://forefire.univ-corse.fr/?page=api/v1/jsdoc/index.html) on the website page. It should be linked to forefire.readthedocs.io.
- Create a "Source Code" section to link the code repository from your website.
- State clearly that "AR Experiments" and/or "YouTube Channel" is to demostrate the further use of the ForeFire simulator API. They are (or are not) part of [ForeFire](https://github.com/forefireAPI/forefire).
- Add a note to the web simulator to state if they are part of ForeFire. If they have a code repository, provide link to access the web simulator if available.
- State that "You can explore the C++ and Python source code on GitHub. The code for web simulator demo can be accessed elsewhere at xxx."

Update your README in the GitHub repository by:
- Use the github "website" field
- State the relationship of "ForeFire Super Console" (up by docker) and [Demo ForeFire API V1](https://forefire.univ-corse.fr/sim/).
