### 2025-08-24 forefireAPI/forefire#102 @xiazeyu
publish python binding package
Please considering publishing your [python binding](https://github.com/forefireAPI/forefire/tree/master/bindings/python) package to [PyPI](https://pypi.org/).

Ref: https://github.com/openjournals/joss-reviews/issues/8680
