### 2025-08-24 forefireAPI/forefire#101 @filippi
hi @xiazeyu thanks
will try to do that this week, this is a small side website to put some demos and links, so you're right, best is to reference GitHub repo which is the main thing now.
regards
