### 2025-08-25 forefireAPI/forefire#103 @xiazeyu
python binding is not installing dependencies
Hello,

I tried the `pip install -e .` inside of the docker container from [Quick Start guide](https://github.com/forefireAPI/forefire/tree/master?tab=readme-ov-file#quick-start-with-docker). However, necessary dependencies like `numpy` and `matplotlib` are not installed.

Please take a look so it installs the dependcies that as well.

Ref: https://github.com/openjournals/joss-reviews/issues/8680
