### 2025-08-25 forefireAPI/forefire#104 @xiazeyu
Unable to build MesoNH, considering building a CI?
Hello,

I understand this might not be in scope of the ForeFire project, but when I'm testing the integration of ForeFire with MesoNH, I failed at building MesoNH. It mentioned something wrong with ld without further explanation of the problem.

Considering you have a dockerfile and full set of test cases, and to better satisify JOSS's "Automated tests" requirement, I recommend to set up a [Continuous integration](https://docs.github.com/en/actions/get-started/continuous-integration) with [GitHub Action](https://docs.github.com/en/actions).

It will be helpful to have to replicate the software with full functionality.

Ref: https://github.com/openjournals/joss-reviews/issues/8680
