### 2025-08-25 forefireAPI/forefire#105 @xiazeyu
Unable to run the real_case example
Hello,

I'm following instructions https://forefire.readthedocs.io/en/latest/getting_started/running_the_example.html. I'm executing in the docker container created from Quick Start.

On [Method 1: Direct Execution](https://forefire.readthedocs.io/en/latest/getting_started/running_the_example.html#direct-execution-command-line-batch-mode), when I executeded the command `forefire -i real_case.ff`, there's no output from the console.

And in the folder, I can't tell which ones are the output files.

On [Method 2: Interactive Console](https://forefire.readthedocs.io/en/latest/getting_started/running_the_example.html#interactive-console), when I executeded the command `include[input=real_case.ff]`, it outputs "wrong input file, check your settings...". That is the same for the web interface.

Could you please have a check, thank you!

Ref https://github.com/openjournals/joss-reviews/issues/8680
