### 2025-08-25 forefireAPI/forefire#106 @xiazeyu
Mention the target audience
Could you please mention the target audience both in README and the paper.

Thank you!

Ref https://github.com/openjournals/joss-reviews/issues/8680
