### 2025-08-25 forefireAPI/forefire#106 @xiazeyu
Sorry. I found that at the end of the documentation. https://forefire.readthedocs.io/en/latest/about/statement_of_need.html
