### 2025-08-25 forefireAPI/forefire#106 @xiazeyu
Recommended promotion to a more prominent position
