### 2025-08-25 forefireAPI/forefire#107 @xiazeyu
simplify redundant parts in statement of need in document
Please consider remove the **Context** part in the [Statement of Need](https://forefire.readthedocs.io/en/latest/about/statement_of_need.html). That part seems redundant and contains potential residual prompt.

Ref https://github.com/openjournals/joss-reviews/issues/8680
