### 2025-08-26 forefireAPI/forefire#105 @xiazeyu
include[real_case.ff] is working. There might be some document issue, but I suggest make it more clear about the expected output of the direct execution method.
