### 2025-08-26 forefireAPI/forefire#104 @filippi
Hi @xiazeyu

Thanks for the comment and advises
CI is already on for part of ForeFire, see https://github.com/forefireAPI/forefire/tree/master/.github/workflows. The MesoNH coupled cases are not part of the standard tests, and actually ForeFire acts as a "plug-in" to MesoNH, to runt its scenarios... that would require big weather data anyway.
Nevertheless, I am ok, it would be great to have it for the coupled version to.. but there are a few problems on the way, first is that MesoNH it is not hosted on GitHub, an that it is a big Fortran code kind of not trivial to have it run in docker .. and for which I would have no control anyway as MesoNH devs do really prefer people to compile it themselves.
Second, MesoNH  devs (very close scientific team) are working on a major GPU/Optimisation release for early next year that is likely to break much stuff and they yet would think more on packaging after that milestone.
Eventually, another thing is that the coupling requires MPI, again some compilation issues not totally trivial that requires mostly the MesoNH team work.. so yet, unable to provide a easy way to run, we will discuss that in the newt MesoNH meeting, until then it will most likely stay the hard way of tar/configure/make, sorry about that.
I asked @antonio-leblanc to look for the small issues on documentation, thanks so much on pointing that, and for all the time you dedicate on this !
