### 2025-08-26 7f2ead0
rm wrong 'input=' instruction from include command #105
