### 2025-08-26 89b78b0
ci: increment version [skip ci]
