### 2025-08-26 8c4c997
for parallel global fire initialisation
