### 2025-08-26 cfad916
improve run example #105
