### 2025-08-26 8cf166b
ci: increment version [skip ci]
