### 2025-08-26 8b71519
ci: increment version [skip ci]
