+++ src/include/Version.h
@@ -1,3 +1,3 @@
 #pragma once
-const char* ff_version = "v2.1.121";
+const char* ff_version = "v2.1.122";
 
