### 2025-08-26 forefireAPI/forefire#105 @antonio-leblanc
Thanks for pointing out, fixed in #110

- rm wrong "input=" arguments
- specified expected outputs
- renamed the page "Running an example" -> "Execution modes: for clarity
