### 2025-08-26 forefireAPI/forefire#105 @xiazeyu
Thank you! That's way much better in my opinion.
