### 2025-08-29 9a3208b
added cmd to dockerfile
