### 2025-08-31 forefireAPI/forefire#101 @filippi
Hi @xiazeyu
I have update the website ! thanks for your comments :-)
https://forefire.univ-corse.fr
Hope it solves the issue !
