### 2025-09-01 forefireAPI/forefire#103 @xiazeyu
I tried to build and run the new docker image. Inside it, I can successfully execute `python3`, but `numpy` and `matplotlib` seems not installed with the python binding. Actually that also happens when I tried to install the binding using `pip install -e .`

Could you please check on that as well?
