### 2025-09-01 forefireAPI/forefire#104 @xiazeyu
Thanks for your replying. That's understandable yet still sad to hear about.

I would assume users are not experties in compiling codes, especially when they are coupled. Have them integrated with each other might be a good future direction. It is also encouraged to have more documtations and examples about how to use forefire with MesoNH for reproducibility.

Personally I can't test the MesoNH integration because I ran into weird errors. But considering forefire is acting as a "plug-in" to MesoNH, I won't block the process because of that.

<summary>Log at compiling</summary>

```
root@6211d16a89f6:~/MNH-V5-7-2/src# make V=1 -j4

make[1]: Entering directory '/root/MNH-V5-7-2/src'

make[1]: '/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/filedepallmaster' is up to date.

make[1]: Leaving directory '/root/MNH-V5-7-2/src'

make[1]: Entering directory '/root/MNH-V5-7-2/src'

make[1]: Nothing to be done for 'objmaster'.

make[1]: Leaving directory '/root/MNH-V5-7-2/src'

make[1]: Entering directory '/root/MNH-V5-7-2/src'

cd /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER ; find . -follow -name '*.o'| xargs ar r /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a

rm -f  /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a

ar: creating /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a

gfortran  -o /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/MESONH /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/spll_mesonh.o   /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib64 -leccodes_f90 -leccodes -lstdc++ -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib64 -lnetcdff -lnetcdf -lhdf5hl_fortran -lhdf5_hl -lhdf5_fortran -lhdf5 -lsz -laec -lz -ldl

gfortran  -o /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/LATLON_TO_XY /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/spll_latlon_to_xy.o   /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib64 -leccodes_f90 -leccodes -lstdc++ -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib64 -lnetcdff -lnetcdf -lhdf5hl_fortran -lhdf5_hl -lhdf5_fortran -lhdf5 -lsz -laec -lz -ldl

gfortran  -o /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/PREP_IDEAL_CASE /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/spll_prep_ideal_case.o   /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib64 -leccodes_f90 -leccodes -lstdc++ -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib64 -lnetcdff -lnetcdf -lhdf5hl_fortran -lhdf5_hl -lhdf5_fortran -lhdf5 -lsz -laec -lz -ldl

gfortran  -o /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/PREP_REAL_CASE /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/spll_prep_real_case.o   /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib64 -leccodes_f90 -leccodes -lstdc++ -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib64 -lnetcdff -lnetcdf -lhdf5hl_fortran -lhdf5_hl -lhdf5_fortran -lhdf5 -lsz -laec -lz -ldl

gfortran  -o /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/PREP_PGD /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/spll_prep_pgd.o   /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib64 -leccodes_f90 -leccodes -lstdc++ -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib64 -lnetcdff -lnetcdf -lhdf5hl_fortran -lhdf5_hl -lhdf5_fortran -lhdf5 -lsz -laec -lz -ldl

gfortran  -o /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/PREP_NEST_PGD /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/spll_prep_nest_pgd.o   /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib64 -leccodes_f90 -leccodes -lstdc++ -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib64 -lnetcdff -lnetcdf -lhdf5hl_fortran -lhdf5_hl -lhdf5_fortran -lhdf5 -lsz -laec -lz -ldl

/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib/libeccodes.a(DataG1ComplexPacking.cc.o): in function `_GLOBAL__sub_I_DataG1ComplexPacking.cc':

/root/MNH-V5-7-2/src/LIB/eccodes-2.41.0-Source/src/eccodes/accessor/DataG1ComplexPacking.cc:13:(.text.startup+0x18): relocation truncated to fit: R_AARCH64_CALL26 against symbol `eccodes::accessor::DataG1ComplexPacking::DataG1ComplexPacking()' defined in .text._ZN7eccodes8accessor20DataG1ComplexPackingC2Ev[_ZN7eccodes8accessor20DataG1ComplexPackingC5Ev] section in /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib/libeccodes.a(DataG1ComplexPacking.cc.o)

/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib/libeccodes.a(step_unit.cc.o): in function `_GLOBAL__sub_I_step_unit.cc':

/root/MNH-V5-7-2/src/LIB/eccodes-2.41.0-Source/src/eccodes/step_unit.cc:15:(.text.startup+0x5c): relocation truncated to fit: R_AARCH64_CALL26 against symbol `std::vector<eccodes::Unit::Value, std::allocator<eccodes::Unit::Value> >::vector(std::initializer_list<eccodes::Unit::Value>, std::allocator<eccodes::Unit::Value> const&)' defined in .text._ZNSt6vectorIN7eccodes4Unit5ValueESaIS2_EEC2ESt16initializer_listIS2_ERKS3_[_ZNSt6vectorIN7eccodes4Unit5ValueESaIS2_EEC5ESt16initializer_listIS2_ERKS3_] section in /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib/libeccodes.a(step_unit.cc.o)

/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib/libeccodes.a(step_unit.cc.o): in function `__static_initialization_and_destruction_0':

/root/MNH-V5-7-2/src/LIB/eccodes-2.41.0-Source/src/eccodes/step_unit.cc:21:(.text.startup+0xb4): relocation truncated to fit: R_AARCH64_CALL26 against symbol `std::vector<eccodes::Unit::Value, std::allocator<eccodes::Unit::Value> >::vector(std::initializer_list<eccodes::Unit::Value>, std::allocator<eccodes::Unit::Value> const&)' defined in .text._ZNSt6vectorIN7eccodes4Unit5ValueESaIS2_EEC2ESt16initializer_listIS2_ERKS3_[_ZNSt6vectorIN7eccodes4Unit5ValueESaIS2_EEC5ESt16initializer_listIS2_ERKS3_] section in /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib/libeccodes.a(step_unit.cc.o)

gfortran  -o /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/SPAWNING /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/spll_spawning.o   /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib64 -leccodes_f90 -leccodes -lstdc++ -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib64 -lnetcdff -lnetcdf -lhdf5hl_fortran -lhdf5_hl -lhdf5_fortran -lhdf5 -lsz -laec -lz -ldl

gfortran  -o /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/DIAG /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/spll_diag.o   /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib64 -leccodes_f90 -leccodes -lstdc++ -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib64 -lnetcdff -lnetcdf -lhdf5hl_fortran -lhdf5_hl -lhdf5_fortran -lhdf5 -lsz -laec -lz -ldl

gfortran  -o /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/PREP_SURFEX /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/spll_prep_surfex.o   /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/ECCODES-2.41.0/lib64 -leccodes_f90 -leccodes -lstdc++ -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib -L/root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/NETCDF-4.6.1/lib64 -lnetcdff -lnetcdf -lhdf5hl_fortran -lhdf5_hl -lhdf5_fortran -lhdf5 -lsz -laec -lz -ldl

collect2: error: ld returned 1 exit status

make[1]: *** [Makefile:444: MESONH] Error 1

make[1]: *** Waiting for unfinished jobs....

rm /root/MNH-V5-7-2/src/dir_obj-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG/MASTER/lib-LXgfortran-R8I4-MNH-V5-7-2-MPIVIDE-DEBUG-MASTER.a

make[1]: Leaving directory '/root/MNH-V5-7-2/src'

make: *** [Makefile:228: progmaster] Error 2
```
