### 2025-09-02 9487d1b
#103 python bindings
