+++ .github/workflows/macos.yml
@@ -3,9 +3,9 @@ name: macOS
 on:
   push:
     branches:
-      - dev
       - brew
-  
+  pull_request:
+    branches: [ "master" ]  
   workflow_dispatch:
 
 jobs:
