### 2025-09-02 forefireAPI/forefire#103 @antonio-leblanc
yes the code is running with some warnings though...

i will check with filippi if maybe there is a better example that can be used on the documentation
