### 2025-09-03 forefireAPI/forefire#103 @filippi
Hi @antonio-leblanc @xiazeyu

The python is kind of supposed to do nothing to show you it is well installed. It is just there so it does not say "no packages named pyforefire".. and that it can create a domain.. then you know it is technically well installed

There is also no 3 liners to perform a sensible simulation of something, to do that, it is best to check in the /tests/python directory
It will still emit a warning (because no fuel tables can be found.. I keep it as default because it is a very common misuse not to use user-defined fuel tables instead of default), but if you want actual simulation, you may change it to do :

import pyforefire as forefire
ff = forefire.ForeFire()
sim_shape = (10000,10000)
ff.execute(f'FireDomain[sw=(0,0,0);ne=({sim_shape[0]},{sim_shape[1]},0);t=0]')
ff.addLayer("propagation","Iso","propagationModel")
ff.execute(f'startFire[loc=({sim_shape[0]/2},{sim_shape[1]/2},0.0)]')
ff.execute("step[dt=1000]")
print(ff.execute("print[]"))
if(False):  # Change to True to export KML
    ff["dumpMode"] = "kml"
    ff.execute("print[circle.kml]")

that will make a nice circle and event create a kml of this circle in Corsica, the default projection area when none is provided.
