### 2025-09-03 15f3107
update pyforefire example with filippi suggestion on #103
