### 2025-09-03 forefireAPI/forefire#103 @antonio-leblanc
Thanks for the example @filippi
I updated the readme of pyforefire accordingly, https://github.com/forefireAPI/forefire/pull/116
