### 2025-09-15 forefireAPI/forefire#117 @xiazeyu
Fuels File
It is mentioned that "the file requires all columns" in the [guide](https://forefire.readthedocs.io/en/latest/user_guide/fuels_file.html).

- Is it possible to set certain values, so the model ignores certain columns if their value is missing or unknown?

- Can you provide some link/references/empirical values for the fuel file? I found some in the test case. Are they from other works?

Ref https://github.com/openjournals/joss-reviews/issues/8680
