### 2025-09-16 forefireAPI/forefire#118 @xiazeyu
Suggestions on CI
Hello,

Here's some suggestions that might make CI better.

- Could you remove unused Workflows in GitHub Action?

- Could you push the built docker image onto [GitHub Packages](https://docs.github.com/en/actions/tutorials/publish-packages/publish-docker-images#publishing-images-to-github-packages)?

- Can you run test scripts for [docker](https://github.com/forefireAPI/forefire/blob/master/.github/workflows/docker.yml) and [macos](https://github.com/forefireAPI/forefire/blob/master/.github/workflows/macos.yml), after their "Check ForeFire version"?

Thanks.

Ref https://github.com/openjournals/joss-reviews/issues/8680
