### 2025-09-16 forefireAPI/forefire#119 @xiazeyu
Software paper comments
1. In Line 29, please add a more recent reference for Flammap.

1. In Line 35, please expand little more about how ForeFire is **adaptable** and **high-performing**. E.g., why it's adaptable, and what features enables it to be high-performing.

1. After Line 35, please mention your relation compared with other works, as well as ForeFire's underlying fire spreading models (list few well-known supported models).

1. After Line 35, please mention who the target audience is.

1. Please state how ForeFire compares with other commonly-used packages.

1. Please cite any dataset that has been pre-compiled into ForeFire.

1. Please consider citing the following works:

    a. https://doi.org/10.3390/fire3030026

    b. https://doi.org/10.1016/j.envsoft.2025.106401

    c. https://doi.org/10.1175/JAMC-D-12-023.1

    d. https://doi.org/10.1016/j.envsoft.2019.104616

Ref https://github.com/openjournals/joss-reviews/issues/8680
