### 2025-09-16 forefireAPI/forefire#119 @filippi
Hi @xiazeyu thanks for comments,
I'll try to do most of it, the thing is that it was there in a  [previous version](https://github.com/forefireAPI/forefire/blob/246d2f936e7caa8e13db1bc6e2c02e1593b66e53/paper/paper.md)  but we have strong issues with paper length to pass. I was more thinking of CELL2FIRE and ELMFIRE that are both used operationally and open sourced, you package PyTorchFire is great, but not yet used in operation, I can mention "other ai codes [your ref]" if you want.
Anyway.. to make all these additions, I would have to remove other parts. Which one are the most important for you ?
