### 2025-09-16 2e98870
fuels and models file #117
