+++ docs/source/user_guide/fuels_and_models.rst
@@ -1,0 +1,101 @@
+.. _userguide-fuels-and-models:
+
+Fuel & Model Parameterization
+=============================
+
+A core feature of ForeFire is its flexibility as a **general Rate of Spread (ROS) model solver**. Rather than being tied to a single fire spread equation, it can use different physical or empirical models depending on the user's needs. This is controlled via the :ref:`propagationModel <param-propagationModel>` parameter.
+
+.. note::
+   **For New Users:** The most common use case for ForeFire is simulating surface wildfires using the standard **Rothermel** model. This model **requires a ``fuels.csv`` file**. If this is your goal, you can skip directly to the section :ref:`The Rothermel / Balbi Fuel File <rothermel-fuel-file>` to learn how to prepare this critical input.
+
+The choice of propagation model is critical because it dictates what physical parameters the simulation requires. This guide explains the available models and how to provide the necessary parameterization for each.
+
+Available Propagation Models
+----------------------------
+
+ForeFire includes several built-in propagation models. Below are the primary models intended for general use. For a complete list of all implemented models, including experimental ones, users can refer to the C++ source code.
+
+**Iso**
+^^^^^^^
+*   **Description:** A simple isotropic model where the fire spreads at a constant rate in all directions. It ignores all environmental factors except for blocking terrain.
+*   **Use Case:** Ideal for simple tests, debugging setups, or scenarios where a constant spread rate is desired.
+*   **Parameterization:** This model **does not** use an external fuels file. The spread speed is set directly in the simulation script via a model-specific parameter.
+*   **Example:**
+    .. code-block:: none
+
+       setParameter[propagationModel=Iso]
+       setParameter[Iso.speed=0.5]  # Sets spread to 0.5 meters/second
+
+**Rothermel**
+^^^^^^^^^^^^^
+*   **Description:** Implements the Rothermel (1972) surface fire spread model, a widely used semi-empirical model in wildfire simulation.
+*   **Use Case:** The standard choice for simulating surface fire spread in various fuel types (grass, shrub, timber litter).
+*   **Parameterization:** This model requires a detailed set of fuel bed parameters. These are provided via a **CSV file**, the format of which is detailed in the section below.
+
+**Balbi** variants (e.g., `BalbiNov2011`, `Balbi2015`)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+*   **Description:** Implements physical models based on the work of Balbi et al. These models take a different approach from Rothermel, often considering factors like flame geometry and radiative heat transfer more explicitly.
+*   **Use Case:** Primarily for research applications, model comparison studies, or scenarios where a more physics-based approach is preferred.
+*   **Parameterization:** These models also source their parameters from the same CSV fuel file format used by Rothermel, though they may use a different subset of the columns for their calculations.
+
+.. note::
+   The relationship between a model and its parameters is key: when a model like `Rothermel` is selected, it informs the simulation engine which properties it needs (e.g., `slope`, `normalWind`, `fuel.Rhod`, etc.). The engine then provides these values, retrieving the `fuel.*` properties from the loaded CSV table.
+
+---
+
+.. _rothermel-fuel-file:
+
+The Rothermel / Balbi Fuel File (`fuels.csv`)
+-----------------------------------------------
+
[…]
