### 2025-09-16 620c158
improve fuel docs #117
