### 2025-09-16 3917a4c
improve basic config for fuels #117
