### 2025-09-16 forefireAPI/forefire#117 @antonio-leblanc
Hi all, thanks again for the feedbacks!

Updated the documentation to address the points discussed here:
- Created a new "Fuel & Model Parameterization" page to clarify that ForeFire is a general ROS model solver.
- The new page now lists the primary propagation models (Iso, Rothermel, Balbi) and explains that the fuels.csv file is a specific requirement for models like Rothermel, not a universal one.
- Added a link to the wildfire_ROS_models repository, as suggested, as a resource for fuel parameter sets.
- Updated the parameter reference and other guides to point to this new, clearer explanation.

I published into another branch inside RTD

maybe @filippi  wanna give a look?

https://forefire.readthedocs.io/en/docs/user_guide/fuels_and_models.html

if Ok we can open PR to dev&master
