### 2025-09-16 397bd3c
add test script to ci for mac-os #118
