+++ .github/workflows/macos.yml
@@ -25,7 +25,7 @@ jobs:
         run: ./bin/forefire -v
 
       - name: Install Python test dependencies
-        run: pip3 install lxml xarray netCDF4
+        run: pip3 install --break-system-packages lxml xarray netCDF4
 
       - name: Run 'runff' Test Script
         run: |
