+++ .github/workflows/macos.yml
@@ -8,7 +8,8 @@ on:
     branches: [ "master" ]  
   workflow_dispatch:
 
-jobs build-native-macos:
+jobs:
+  build-native-macos:
     runs-on: macos-latest
     steps:
       - name: Checkout repository
