### 2025-09-16 dbc900b
add test to docker #118
