### 2025-09-16 forefireAPI/forefire#118 @antonio-leblanc
MacOs CI ok https://github.com/forefireAPI/forefire/actions/runs/17774763753/job/50519607189
