### 2025-09-16 forefireAPI/forefire#118 @antonio-leblanc
Docker ci testing oK

https://github.com/forefireAPI/forefire/actions/runs/17775281120
