### 2025-09-16 2c337fa
organizing devops tools not needed now #118
