### 2025-09-16 forefireAPI/forefire#117 @filippi
Thanks @xiazeyu
