### 2025-09-16 forefireAPI/forefire#119 @xiazeyu
Hello @filippi,

1, 3, 4, 7 are most important to me.

Please try to cite some commonly used simulators for operational use, and some more experimental models like PyTorchFire. One sentence comparison for each will be enough to show why people may use your simulator.

Please also mention the target audience. It's a requirement for JOSS.
