+++ paper/paper.bib
@@ -22,6 +22,34 @@ title = {{FARSITE: Fire Area Simulator-model development and evaluation}},
 url = {http://www.treesearch.fs.fed.us/pubs/4617},
 year = {1998}
 }
+
+@misc{Finney2023FlamMap,
+  author       = {Finney, Mark A. and Brittain, S. and Seli, R. C. and McHugh, C. W. and Gangi, L.},
+  title        = {{FlamMap: Fire Mapping and Analysis System (Version 6.2) [Software]}},
+  year         = {2023},
+  howpublished = {\url{https://www.firelab.org/project/flammap}},
+  note         = {Available from U.S. Forest Service Fire Sciences Laboratory}
+}
+@article{XIA2025106401,
+title = {PyTorchFire: A GPU-accelerated wildfire simulator with Differentiable Cellular Automata},
+journal = {Environmental Modelling & Software},
+volume = {188},
+pages = {106401},
+year = {2025},
+issn = {1364-8152},
+doi = {https://doi.org/10.1016/j.envsoft.2025.106401},
+url = {https://www.sciencedirect.com/science/article/pii/S1364815225000854},
+author = {Zeyu Xia and Sibo Cheng},
+keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based software, Parallel computing techniques, GPU-acceleration},
+}
+@book{Scott2005,
+  title = {Standard fire behavior fuel models: a comprehensive set for use with Rothermel’s surface fire spread model},
+  url = {http://dx.doi.org/10.2737/RMRS-GTR-153},
+  DOI = {10.2737/rmrs-gtr-153},
+  institution = {U.S. Department of Agriculture,  Forest Service,  Rocky Mountain Research Station},
+  author = {Scott,  Joe H. and Burgan,  Robert E.},
+  year = {2005}
+}
 @article{mell2007,
   title = {A physics-based approach to modelling grassland fires},
   volume = {16},
@@ -151,6 +179,7 @@ year = {1998}
   month = aug,
   pages = {629–646}
 }
+
 @article{santoni2011,
   title = {Wildland Fire Behaviour Case Studies and Fuel Models for Landscape‐Scale Fire Modeling},
   volume = {2011},
