### 2025-09-16 forefireAPI/forefire#119 @filippi
Last compiled paper :

https://raw.githubusercontent.com/openjournals/joss-papers/joss.08680/joss.08680/10.21105.joss.08680.pdf

Thank again for comments
