### 2025-09-16 forefireAPI/forefire#119 @xiazeyu
LGTM. Thanks to your efforts.
