+++ paper/paper.bib
@@ -37,7 +37,7 @@ volume = {188},
 pages = {106401},
 year = {2025},
 issn = {1364-8152},
-doi = {https://doi.org/10.1016/j.envsoft.2025.106401},
+doi = {10.1016/j.envsoft.2025.106401},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815225000854},
 author = {Zeyu Xia and Sibo Cheng},
 keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based software, Parallel computing techniques, GPU-acceleration},
