### 2025-09-16 forefireAPI/forefire#118 @xiazeyu
@antonio-leblanc The GitHub packages page shows 404 Not Found.

Also, the CI seems broken. https://github.com/forefireAPI/forefire/actions/runs/17779578287/job/50535568281

I believe there's something wrong so it is testing on `bash:latest` instead.
