### 2025-09-16 6c458cf
fix docker ci when no tags are created
