### 2025-09-17 forefireAPI/forefire#118 @antonio-leblanc
Thanks @xiazeyu  the package was private that's why...

I don't have the permissions to set the package public
Could you do it please @filippi

https://github.com/orgs/forefireAPI/packages/container/forefire/settings
