### 2025-09-17 49150f4
adding docker badge to readme  #118 + reorder badges
