### 2025-09-17 forefireAPI/firefront#125 @iperezx
Getting Started README example doesn't display data.
Reference: https://github.com/openjournals/joss-reviews/issues/8680
I am following the [Quick Start with Docker section](https://github.com/forefireAPI/forefire?tab=readme-ov-file#quick-start-with-docker)
It built and I am able to see the web UI.
I am not sure if I am erroring out on the last step (Run your first simulation) but I don't see anything being populated on web UI.

This is what I see in the logs:
```
forefire> listenHTTP[]
ForeFire HTTP command server listening at http://localhost:8000
To stop the server, press Ctrl-C or type quit in the ForeFire shell.
forefire> File not found: js/leaflet-velocity.css
File not found: img/forefire_logo.svg
File not found: js/rosacewind.js
File not found: js/forefireGUI.js
File not found: js/leaflet-velocity.js
File not found: img/ff_logo.png
File not found: favicon.ico
File not found: apple-touch-icon-precomposed.png
File not found: apple-touch-icon.png
NetCDF: Unknown file format
file: ncFile.cpp  line:88
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
File not found: js/leaflet-velocity.css
```
