### 2025-09-17 forefireAPI/forefire#126 @iperezx
Slow UI boot up
Reference: https://github.com/openjournals/joss-reviews/issues/8680

I am fine with ignoring this one, but wanted to bring this up since it will help with high adoption of your platform.
When the user first goes to localhost:8000 (if they start from the getting started part), the boot up for the UI takes a while.
Is it trying to build everything from scratch?
Noticed that this doesn't happened on the second time you go to localhost:8000.

Some recommendations:
- Why not build the UI assets during the docker build process?
