### 2025-09-17 forefireAPI/forefire#125 @antonio-leblanc
Hi @iperezx

You need to type include[real_case.ff]
and send it, this makes forefire simulate the event. Indeed nothing appears at this moment

But after that you need to press the button "Refresh Map" the simulation appears..

But i understand this can be a bit tricky so we'll think of a way to explain better on readme or maybe change the button color to make it stand more
