### 2025-09-17 forefireAPI/forefire#125 @filippi
Hi @iperezx and thanks @antonio-leblanc

Yes.. the GUI is more of a shortcut to an interpreter, you may even actually launch include and the listenHTTP,
Actually I kept the fact to "refresh map" because "refresh map" is actually also a command.
so that the interface itself is like a way to understand the philosophy (everything is a command/instruction).. somehow problem is that quick start is not totally possible... would need a bit more..

Maybe we can change the "quick start" to state:
"everything is command in ForeFire, open browser, then click "include" button to pre-write the include[realcase.ff] command, then "send" then "refresh" to launch the "print[output.json]" command that is linked to update the display.
what do you think @iperezx  ??

also, the html interface is highly customizable, it is complimentary to ForeFire, more a showcase of what would be possible to do the we'll have a nice community making plenty of interfaces in 2D/3D.. ensembles... AI..
