### 2025-09-17 forefireAPI/forefire#126 @filippi
Hi @iperezx
I can't really reproduce, never had the issue
, There is some chance that the docker image is serving a bit slow depending on how it is configured, once js is pre-compiled and cached better, and network checks done (given it opens some port, you may have also some security service ?)..
Do we need to add a comment in readme ?
