### 2025-09-17 7302697
dont print File not found if FOREFIRE home var exists #125
