### 2025-09-18 b5fac7b
organize script into other file + add imprort to top of html #125
