+++ tools/htdocs/js/forefireGUI.js
@@ -15,7 +15,7 @@ const commands = {
     "setParameter": "setParameter[param=value]",
     "setParameters": "setParameters[param1=val1;param2=val2]",
     "getParameter": "getParameter[paramNames]",
-    "include": "include[run.ff]",
+    "include": "include[real_case.ff]",
     "loadData": "loadData[data.nc;2024-12-13T15:41:33Z]",
     "clear": "clear[]",
     "systemExec": "systemExec[ls]",
