### 2025-09-18 16cd9bc
rm event listener refreshMap that was being called twice
