### 2025-09-18 7590f79
better quick start docs + buttons #125
