### 2025-09-18 290623a
better quick start gui #125
