### 2025-09-18 c1fdfc1
Improve Badges on Readme + Remove "File Not Found" errors on terminal if file actually found (#127)
* adding docker badge to readme  #118 + reorder badges

* dont print File not found if FOREFIRE home var exists #125

* split css from html

* organize script into other file + add imprort to top of html #125

* defauls "include": "include[real_case.ff]" #125

* include as 1st button #125

* rm event listener refreshMap that was being called twice

* better quick start docs + buttons #125

* better quick start gui #125

---------
