### 2025-09-18 forefireAPI/forefire#126 @iperezx
I was able to produce it if I use a private/incognito mode. I can also avoid it if I never leave the web page where I first started it.
For reference my browser is clocking it at 29 seconds.

I am able to serve some of our internal platforms (with Docker) within the typical millisecond range for UIs (assuming that it was already built).

Again you can just ignore this but it does help.
