### 2025-09-18 c457e4c
image closer to the explanation of where user needs to click #125
