### 2025-09-18 forefireAPI/forefire#125 @antonio-leblanc
@iperezx
please can you provide more info on the setup you are running?
- what is the OS
- Are you using the docker right?
- What version of forefire appears when launching

it will help on the debug here thanks
