### 2025-09-19 forefireAPI/forefire#125 @antonio-leblanc
Hi @iperezx
As filippi said this looks like a problem with LFS...
I will add to the documentation that users need to install an extension (probably only for mac, i will confirm later)

But in a nutshell
- *Git LFS* (Large File Storage) is a Git extension designed to handle large binary files efficiently without bloating the core Git repository.
- Instead of committing the large file directly, LFS commits a tiny text pointer file into the Git repository. This pointer contains a unique ID for the large file. The actual large file is uploaded to a separate, dedicated storage server (like GitHub LFS).
- The data.nc file is a large data file, and your local machine downloaded a small text pointer instead of the actual file

Since i don't use mac i was getting gemini to explain me the suggested steps...

Could you try downloading and installing lfs please?

```
brew install git-lfs
git lfs install
```

Then DELETE the current forefire repository you cloned on your machine

and after that clone it again
```
git clone https://github.com/forefireAPI/forefire.git
```

And normally now you should have the data.nc file downloaded....

if those steps work for you i will update the docs...

Sorry i cannot test because i use windows + WSL and this problem does not happen

Thanks a lot
