### 2025-09-19 forefireAPI/forefire#125 @filippi
Hi @iperezx

It is very likely there is an issue with git LFS when you cloned, can you check ?
