### 2025-09-20 forefireAPI/forefire#125 @iperezx
I can reproduce the same error on Windows + WSL.

It might be a git LFS issue. Weird that a file of size 4MB you need to use git LFS.

Any particular reason why you are storing data files in the repo?

Traditionally, data files (that are big) are kept somewhere else and in the build stage you either do `wget`, `curl`, etc.

It helps with trying to keep the repo light weight.
