### 2025-09-20 forefireAPI/forefire#126 @iperezx
On Windows + WSL + Chrome, it seems like there is no lag to see the UI.
