### 2025-09-20 forefireAPI/forefire#126 @filippi
Cool @iperezx thanks, on my m4 no problem either..
I'll see if I can reproduce on other machine
