### 2025-09-20 forefireAPI/forefire#125 @iperezx
Oh you purposely added the file as a git LFS pointer file?
I thought you added the file and then github decided to add the pointer file. This was my experience before.
Maybe include that part in your docs that you also need have git LFS. I think its not default of git.

Dont think its misused. Just that they have limits per file: https://docs.github.com/en/repositories/working-with-files/managing-large-files/about-git-large-file-storage#about-git-large-file-storage
