### 2025-09-20 forefireAPI/forefire#125 @filippi
Hi @iperezx

Yes, purposely, didn't realised it wasn't that standard
It is smallish big files for tests, can't really put it elsewhere more convenient in fact, I know about GitHub FS limit, not big deal yet...
Thanks
Until then you can always download the raw file from github.com site. It should work
Regards
