### 2025-09-20 forefireAPI/forefire#125 @filippi
Also @iperezx , can you please think to check the boxes of your review form for those you think are OK, and close issues if you think it is good, so we can track what may be left to do
Thank so much for your help
Regards
