### 2025-09-21 forefireAPI/forefire#125 @iperezx
I don't think Git LFS is part of the standard Git distribution. Github provides instructions to add it here:[ ](https://docs.github.com/en/repositories/working-with-files/managing-large-files/installing-git-large-file-storage?utm_source=chatgpt.com&platform=windows)

I think if you provide that assumption in your README, I can close this one out.
