### 2025-09-21 forefireAPI/forefire#125 @filippi
Hi @iperezx — thanks for catching that!
I just added a note to the README explaining that the demo datasets in tests/runff/ are stored with Git LFS. If Git LFS isn’t installed, the clone will only have the pointer file, so you can either install Git LFS and run git lfs pull, or download the dataset directly from the GitHub web UI and drop it back into the same folder. These files are only there to support the bundled test examples; regular runs should bring their own inputs. Let me know if you still hit issues on Windows/WSL after that.
