### 2025-09-21 forefireAPI/forefire#125 @iperezx
Now that I have the `data.nc` file.

I get the following error:
```
forefire >include[real_case.ff]
wrong input file, check your settings...
```

This is what I see from ncinfo:
```
root@494abedad0e8:/forefire/tests/runff# ncinfo data.nc
<class 'netCDF4.Dataset'>
root group (NETCDF4 data model, file format HDF5):
    dimensions(sizes): ft(1), fz(1), fy(6400), fx(6400), wind_dimensions(2), wind_directions(2), wind_rows(160), wind_columns(160), nt(1), nz(1), ny(640), nx(640)
    variables(dimensions): <class 'str'> domain(), int16 fuel(ft, fz, fy, fx), float32 wind(wind_dimensions, wind_directions, wind_rows, wind_columns), int16 altitude(nt, nz, ny, nx)
    groups:
```
