### 2025-09-21 forefireAPI/forefire#125 @iperezx
Trying again and I am not seeing the error again in the UI logs but I don't see any indication that the simulation has completed.

I only see the following:
```
forefire >include[real_case.ff]
```

In the README, I see this:
```
Step 2: View the result. After the command finishes, ....
```

How would I know that the command finished? Is there any indication that the simulation completes?

I don't see anything in the logs from the CLI tool.
