### 2025-09-21 forefireAPI/forefire#126 @iperezx
No problem. You can close it. You now have a record of it so if someone encounters it they can see it on the closed issues.
