### 2025-09-21 forefireAPI/forefire#125 @filippi
also @iperezx
basically by "see the results" I would have expected the user to type >print[]

simulation has actually not "completed" in the sense that while in CLI you are still interacting with it.. you may also write :
step[dt=2400]
print[]
and it'll print 40 mins after
a big point of these examples is to get user getting to understand the philosophy of FF sim interaction

thanks
