### 2025-09-21 forefireAPI/forefire#126 @filippi
Thanks @iperezx
