### 2025-09-26 #8680 @iperezx
I think I am still in this stage of the review:
```
- Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
- Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
```
