### 2025-09-27 forefireAPI/forefire#125 @filippi
Dear @iperezx
I may not understand well, do you suggest adding a .github/workflows/readme.md in which we expose the same process ?
What documentation files do we need to touch to solve the issue ? thanks a lot
