### 2025-09-28 forefireAPI/forefire#125 @iperezx
I am only suggesting to have another test in the github actions because so far the section of [Quick Start with Docker](https://github.com/forefireAPI/forefire/tree/master?tab=readme-ov-file#quick-start-with-docker), has not been a quick start experience.

I noticed that this workflow does test most of the parts of the Quick Start section: https://github.com/forefireAPI/forefire/blob/master/.github/workflows/docker.yml

But dont see anything testing step 5 and the other steps that the user has to perform on the UI.

What do you guys think? Definitely push back if it is too much work.
