### 2025-09-28 forefireAPI/forefire#125 @antonio-leblanc
Dear @iperezx
Sorry that this was not a quick start,

Indeed this is a GIT LFS issue., and we cannot test this in the CI/CD since the ci/cd have the lfs enabled
They have the settings `lfs: true`
- Linux https://github.com/forefireAPI/forefire/blob/master/.github/workflows/main.yml#L19
- Docker https://github.com/forefireAPI/forefire/blob/master/.github/workflows/docker.yml#L26
- Mac os https://github.com/forefireAPI/forefire/blob/master/.github/workflows/macos.yml#L18

FIlippi has added this to the readme
"
Sample data and Git LFS
The demo datasets bundled under tests/runff/ are stored with Git LFS because they include several megabytes of raster data that we only use in the quick-start examples and regression tests. Make sure Git LFS is installed before cloning; otherwise Git will pull pointer files only. If that happens, download the dataset directly from the GitHub web interface and drop it back into the expected folder before running the examples. This data is only provided for the bundled test scenarios.
"

Hoever what we can do to make this even clearer is to explicity tell the user to check the size of the data.nc as one of the quick start steps

Today the quick start is
1. Clone the repository
2. Build Docker
3. Run Docker
4. Other steps that test forefire

And we can edit to

1. Clone the repo
2. *Check that data.nc is a big file. If not install git lfs and clone it again*
3. Build docker
4. Run docker
5. Other steps that test forefire

I think this might avoid users to not download the file

I can do this next week @filippi
