### 2025-09-29 forefireAPI/forefire#125 @filippi
Dear @iperezx, sorry to read it has not been smooth... and it is just a problem with LFS and a few lines documentation issue that we are not able to resolve well enough so you can actually close it.

Hi @antonio-leblanc, that will be nice as of me.
@iperezx what solution would fit you to solve the issue :
- make it even clearer that LFS is a dependency (write it very big), with an alternative to download directly (well explained in quick start). As
- we can also make a test cases that do not requires a nc file or git lfs, but it will not be on real landscape.. and i do not like that very much.

because it has bindings with other codes and interpreters (python, fortran), and it is also a scripting language, it is likely that you'll encounter other issues with your install (especially - and this should not be a requirement for the publication - if you want to acheive optional coupling with third party open code Meso-NH for which we have little control over).
I'm therefore actually very worried the review will take a very very long time, if you think the same @iperezx, can you please notify us so we can stat to ask more contributors for solving issues.

Please also consider that, while used by many agencies and companies, this is not a commercial software, and we are not supported by any commercial funding, if it is open and that we want to publish into JOSS it is actually to help build a community to make it (and documentation) better.

Thanks,
