### 2025-09-29 bf5bc60
add check script for .nc data #125
