### 2025-09-29 a87be32
update readme and ci with lfs test #125
