### 2025-10-06 forefireAPI/forefire#125 @filippi
Hi @iperezx did you manage to have it running ?
If we make a hard download, would it be enough so we can close the issue ?
tkanks a lot
