### 2025-10-17 b60f460
Replace local video link with GitHub asset link
Updated video link for Docker quick start instructions.
