+++ README.md
@@ -51,22 +51,20 @@ The easiest way to get started is often using Docker and the interactive console
 <video src="https://github.com/user-attachments/assets/e257fa8c-5880-4b96-a671-5e3af576be48" width="600" autoplay loop muted playsinline></video>
 
 
-#### Windows note
-
-ForeFire targets Unix-like systems (Linux/macOS). On Windows, please use Docker Desktop or WSL2 to run the quick-start examples.
-
 1. Clone the repository
     
     ``` bash
     # Clone the repository
     git clone https://github.com/forefireAPI/forefire.git
     cd forefire
+
     ```
 
 2. Verify Git LFS sample data (required for the quick-start examples)
 
     ```bash
     bash tools/preflight/verify-lfs.sh
+    
     ```
     If this fails (e.g., `tests/runff/data.nc` is ~1 KB), install Git LFS and reclone the repo, or download the file from the GitHub web UI into `tests/runff/`.
 
@@ -87,14 +85,16 @@ ForeFire targets Unix-like systems (Linux/macOS). On Windows, please use Docker 
 
     # start the forefire console with the command
     forefire
+
     ```
 
-6. Inside the console launch an http server with listenHttp[] command
+6. Inside the console launch an http server with listenHTTP[] command
+    ```bash
+    listenHTTP[]
+    ````
+    the output should be :
 
     ```bash
-    forefire> listenHTTP[]
-
-    # the output should be
     >> ForeFire HTTP command server listening at http://localhost:8000
     ```
 
