### 2025-10-17 forefireAPI/forefire#125 @philipcardiff
Hi @filippi and @iperezx,

I can see that this has become a time-consuming issue; nonetheless, being able to install the software and run the first tutorial is obviously a critical aspect of the review.

To help the process, I have tried to install ForeFire and run the first tutorial.

I followed the steps at https://forefire.readthedocs.io/en/latest/getting_started/quickstart.html; I installed it on my headless Ubuntu server and connected to the Web UI via a port forward to my macOS laptop. I can access and view the interface but I get an error when I run ` include[real_case.ff]`. Here are the steps I followed on my Ubuntu server:
```
git clone https://github.com/forefireAPI/forefire.git
cd forefire
docker build . -t forefire:latest
docker run -d --rm --name forefire_srv -p 127.0.0.1:8000:8000 forefire:latest sleep infinity
docker exec -it forefire_srv bash
cd tests/runff
forefire
listenHTTP[]
```
I then set up port forwarding on my laptop, I run `ssh -N -L 8000:localhost:8000 <my_user_name>@<server_address>`.
Then when I run `include[real_case.ff]` on my ubuntu server, I get:
```
forefire> include[real_case.ff]
NetCDF: Unknown file format
file: ncFile.cpp  line:88
Problem Loading data from NetCDF file /forefire/tests/runff/./data.nc: NetCDF: Unknown file format
file: ncFile.cpp  line:88
FireDomain[sw=(1.23457e+09,1.23457e+09,1.23457e+09);ne=(2.46914e+09,2.46914e+09,2.46914e+09);t=1.23457e+09]
Exception in creating a_time: std::bad_alloc
```
