### 2025-10-17 5d6e7c9
init entrypoint script #125
