+++ README.md
@@ -60,26 +60,18 @@ The easiest way to get started is often using Docker and the interactive console
 
     ```
 
-2. Verify Git LFS sample data (required for the quick-start examples)
-
-    ```bash
-    bash tools/preflight/verify-lfs.sh
-    
-    ```
-    If this fails (e.g., `tests/runff/data.nc` is ~1 KB), install Git LFS and reclone the repo, or download the file from the GitHub web UI into `tests/runff/`.
-
-3. Build the Docker image 
+2. Build the Docker image 
 
     ```bash
     docker build . -t forefire:latest
     ```
 
-4. Run the container interactively
+5. Run the container interactively
 
     ```bash
     docker run -it --rm -p 8000:8000 --name ff_interactive forefire
     ```
-5. Inside the container navigate to test directory and launch the forefire console:
+6. Inside the container navigate to test directory and launch the forefire console:
     ```bash
     cd tests/runff
 
@@ -88,8 +80,7 @@ The easiest way to get started is often using Docker and the interactive console
 
     ```
 
-6. Inside the console launch an http server with listenHTTP[] command
-    ```bash
+67   ```bash
     listenHTTP[]
     ````
     the output should be :
@@ -102,9 +93,13 @@ The easiest way to get started is often using Docker and the interactive console
 
 7. Run your first simulation
     
-    In ForeFire, running a simulation and viewing the result are separate commands. The UI guides you through this process.
-    - **Step 1: Run the simulation script.** In the command input box, type `include[real_case.ff]` and click the **`Send`** button. The simulation will run on the server.
-    - **Step 2: View the result.** After the command finishes, click the **`Refresh Map`** button to load the simulation results onto the map.
+    In ForeFire (web or on console are equivalent), running a simulation and viewing the result are separate commands. The UI guides you through this process.
+    - **Step 1: Run the simulation script.** In HTTP Interface click the **`include`** button or type `include[real_case.ff]` in command input box, and click the **`Send`** button. 
+    ou can also run the same command directly in the interactive console if you prefer py typing `include[real_case.ff]` and press enter.
+    The scripts executes a simulation by loading data, starting fires, applying wind triggers, and running the simulation for a specified duration.
+
+    - **Step 2: View the result.** After the command finishes, click the **`Refresh Map`** button to display the simulation results onto the web map.
+    - **Step 3 (optional): iterate more.** you can continue the simulation, command `include[real_case.ff]` the **`Refresh Map`** button to display the simulation results onto the web map.
     
     ![ForeFire Web UI showing a simulation example](docs/source/_static/images/gui_real_case_ff.jpg)
     
