+++ Dockerfile
@@ -19,8 +19,12 @@ RUN pip3 install --no-cache-dir lxml xarray netCDF4
 WORKDIR /forefire
 ENV FOREFIREHOME=/forefire
 
-# we could only copy src, cmakelists.txt and cmake-build.sh
-COPY . .
+# Copy build configuration first (rarely changes)
+COPY CMakeLists.txt cmake-build.sh LICENSE ./
+
+# Copy source code and tools (changes more frequently)
+COPY src/ ./src/
+COPY tools/ ./tools/
 
 # Build and install the ForeFire C++ library
 RUN sh cmake-build.sh
@@ -29,11 +33,16 @@ RUN sh cmake-build.sh
 RUN cp /forefire/bin/forefire /usr/local/bin/
 
 # Use pip to install the Python bindings
+COPY bindings/ ./bindings/
 RUN pip3 install ./bindings/python
 
-# Copy and set up the entrypoint script
-COPY scripts/entrypoint.sh /usr/local/bin/entrypoint.sh
-RUN chmod +x /usr/local/bin/entrypoint.sh
+# Copy everything else (changes most frequently - docs, tests, etc.)
+COPY . .
 
-# Set the entrypoint for automatic demo startup
-ENTRYPOINT ["/usr/local/bin/entrypoint.sh"]
+# WE COULD USE A CUSTOM ENTRYPOINT SCRIPT TO START THE HTTP SERVER AND RUN THE DEMO SIMULATION
+# COPY tools/devops/entrypoint.sh /usr/local/bin/entrypoint.sh
+# RUN chmod +x /usr/local/bin/entrypoint.sh
+# ENTRYPOINT ["/usr/local/bin/entrypoint.sh"]
+
+# Set the entrypoint to bash for interactive sessions
+CMD ["bash"]
