### 2025-10-17 8ca797c
move forefire executable code to /app
