### 2025-10-17 c16cf42
[docker] Faster Docker Rebuilds (#134)
* init entrypoint script #125

* docker CP is more optimized
