### 2025-10-23 forefireAPI/forefire#125 @filippi
Thanks a lot @xiazeyu for your help again

Right now, a ff cannot trigger a HTML DOM event, it would require a timeout, that can be more or less long, hard to know, something we may do is having the "refresh map" timeout checked by default, but I may be slow on smaller machines. We'll make different tests but for a newer version, everything we tested feels a bit clumsy interactions. We'll find something with a new version as we would really like this one out and frozen to start building on.

I'm OK with you for small files don't need LFS, but these are binary files and each time we change version whole file may change even if it changes a byte or 2 (version Nr), so it'll make large commit (whole file). Thanks to the problems raised by @iperezx we have implemented a large number of workarounds that I hope will help overcoming the problem.
Thanks again
Regards
