### 2025-10-24 #8680 @kthyng
@philipcardiff Ok I understand. It sounds like we could add something into the docs as guidelines for how much time is enough time, or to set an expectation for when it is too much. I suggest you bring this up as a discussion topic internally so we can gather thoughts and then get something into the docs afterward. It is a good point, thanks all.
