### 2025-10-26 a570fd3
[docker] Faster docker experience 2 (#135)
* init entrypoint script #125

* docker CP is more optimized

* move forefire executable code to /app

* final docker

* comment
