+++ tests/python/farsite_flat.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
 """
 Created on Thu Jul 18 10:23:07 2024
+Polished on Wed OCt 29 08:35:00 2025
 
 @author: Batti Filippi
+@co-author: Fernando Veiga
 """
 
 import numpy as np
@@ -17,12 +19,9 @@ import matplotlib.pyplot as plt
 import matplotlib.patches as mpatches
 import matplotlib.patches as patches
 import numpy as np
-from matplotlib import cm
+import matplotlib
 import math
 import datetime
-from pyproj import Proj, transform
-
-
 import pyforefire.helpers as ffhelpers
 
 def plot_test(pathes, myExtents, bg_array=None):
@@ -60,7 +59,7 @@ def plot_test(pathes, myExtents, bg_array=None):
     ax.grid(True)
 
     # Create a colormap instance from inferno and determine colors for each path
-    cmap = cm.get_cmap('inferno', len(pathes))
+    cmap = matplotlib.colormaps['inferno']
     
     for idx, path in enumerate(pathes):
         # Use a different color for each path
@@ -71,7 +70,7 @@ def plot_test(pathes, myExtents, bg_array=None):
     ax.axis('equal')
     ax.set_xlim(xmin, xmax)
     ax.set_ylim(ymin, ymax)
-    plt.show()
+    #plt.show()
 
 def load_weather_file(file_path, default_step_size=1800):
     """
@@ -92,6 +91,7 @@ def load_weather_file(file_path, default_step_size=1800):
         for line in f:
             if line.strip().startswith("Year"):
                 header = line.strip().split()
+                print(header)
                 break
         
         # Read the remaining lines as data rows.
@@ -376,19 +376,19 @@ def plot_landscape_with_header(lcp_data):
     plt.tight_layout()
     plt.show()
 
+## Simulation Parameters ##
+nb_steps = 3               # The number of step the simulation will execute
+step_size =   1000              # The duration (in seconds) between each step
 
 # Initialize pyforefire module
 ff = pyff.ForeFire()
-mylcpfile = "/Users/filippi_j/soft/tests/farsite/examples/flatland/Inputs/a_lcpFiles/flatland.lcp"
-mylcpfile = "/Users/filippi_j/soft/tests/farsite/examples/Panther/cust/input/238648_cust_fm94.lcp"
-mylcpfile = "/Users/filippi_j/soft/firefront/local/tests/python/ZHR_37.lcp"
-mylWeatherFile = "/Users/filippi_j/soft/firefront/local/tests/python/weather_MNH_41_69-1_90.wxs"
+mylcpfile = "./flatland.lcp"
+mylWeatherFile = './flatland_3mph0deg7hr.raws'
 a = readLCPFile(mylcpfile)
 
 #plot_landscape_with_header(a)
 
 ff["propagationModel"] = "Farsite"
-pmodel = ff["propagationModel"]
 ff["fuelsTable"] = "STDfarsiteFuelsTable"
 ff["FarsiteLoggerCSVPath"] = "FF_FS4roslog_2.csv"
 
@@ -398,26 +398,14 @@ ff["moistures.tens"] =  0.04
 ff["moistures.livew"] =  1
 ff["moistures.hundreds"] =  0.06
 
-
-ff["moistures.ones"] = 0.06
-ff["moistures.liveh"] =  0.7
-ff["moistures.tens"] =  0.07
-ff["moistures.livew"] =  1
[…]
