+++ tests/python/readme.md
@@ -1,0 +1,5 @@
+# *Python-basesd* Test Suite
+
+## 1. farsite_flat.py
+It runs a simulation based in Farsite software for a north-wind of 3mph.
+Required input files can be downloaded from: https://universidadevigo-my.sharepoint.com/:f:/g/personal/fernando_veiga_uvigo_gal/Eo5vDajmvf5LmCm58XCWsl0Bm08ya3NK499nP7LHitgeFA?e=4Mo4i7
