+++ tests/python/README.md
@@ -3,5 +3,5 @@
 ## 1. farsite_flat.py
 It runs a simulation based in Farsite software for a north-wind of 3mph.
 
-Required input files can be downloaded from: 
-https://universidadevigo-my.sharepoint.com/:f:/g/personal/fernando_veiga_uvigo_gal/Eo5vDajmvf5LmCm58XCWsl0Bm08ya3NK499nP7LHitgeFA?e=4Mo4i7
+Required the input .lcp file can be downloaded from:
+https://github.com/mbedward/farsite/raw/refs/heads/master/examples/flatland/Inputs/a_lcpFiles/flatland.lcp
