+++ tests/python/README.md
@@ -1,0 +1,7 @@
+# *Python-basesd* Test Suite
+
+## 1. farsite_flat.py
+It runs a simulation based in Farsite software for a north-wind of 3mph.
+
+Required the input .lcp file can be downloaded from:
+https://github.com/mbedward/farsite/raw/refs/heads/master/examples/flatland/Inputs/a_lcpFiles/flatland.lcp
