### 2025-11-03 #8680 @philipcardiff
> Hi! I am just writing to let you know that I have no conflict of interest with any of the people involved in the project. I have just started using and iterating for forefire and would be happy to review in any case a third reviewer would be needed, mostly for the python-based interface, in which I intend to keep working and actively improving.
>
> Best!

Thanks @fveigauvigo, much appreciated. I will add you as a reviewer.
