### 2025-11-27 81c60ee
[JOSS] dev to master before release (#139)
* Replace local video link with GitHub asset link

Updated video link for Docker quick start instructions.

* [docker] Faster Docker Rebuilds (#134)

* init entrypoint script #125

* docker CP is more optimized

* [docker] Faster docker experience 2 (#135)

* init entrypoint script #125

* docker CP is more optimized

* move forefire executable code to /app

* final docker

* comment

* orcid for antonio

---------

Co-authored-by: Jean-Baptiste Filippi <batti.filippi@gmail.com>
