+++ paper/paper.md
@@ -92,7 +92,7 @@ Wildfire forecasting is both an active research area and an important need for d
 Wildfire modeling tools have historically been split between **complex combustion research models** and **streamlined operational tools**, each with distinct limitations. 
 Computational combustion and fluid dynamics (CFD) based models (e.g., FIRETEC [@linn2005] or WFDS [@mell2007]) are highly computationally intensive and yet unable to provide large wildfire forecasts faster than real time. 
 Atmospheric coupled codes, such as WRF/SFire [@mandel2011] must be run within an atmospheric model and require a large amount of processing power and data.
-Operational wildfire simulators, such as widely used Farsite [@finney1998] (now Flammap[@Finney2023FlamMap]), or Canadian Prometheus [@garcia2008], are able to simulate fire fronts spanning tens of kilometers in a matter of seconds, but have definite built-in modeling assumptions and are distributed as compiled software with graphical interfaces with limited scriptability. Other open source libraries are ElmFire [@lautenberger2013] or Cell2Fire [@pais2021] that are tied to a single spread models and do not include scripting language, or deep learning based [@XIA2025106401]. 
+Operational wildfire simulators, such as widely used Farsite [@finney1998] (now Flammap [@Finney2023FlamMap]), or Canadian Prometheus [@garcia2008], are able to simulate fire fronts spanning tens of kilometers in a matter of seconds, but have definite built-in modeling assumptions and are distributed as compiled software with graphical interfaces with limited scriptability. Other open source libraries are ElmFire [@lautenberger2013] or Cell2Fire [@pais2021] that are tied to a single spread models and do not include scripting language, or deep learning based [@XIA2025106401]. 
 
 ForeFire was developed as a community tool to fill the gap between highly complex customizable models and more rigid operational tools: a **unified** wildfire simulator that is both **adaptable** (highly scriptable with multiple bindings) and **high-performing** (discrete‑event‑driven simulation with dynamic mesh allows to concentrate computation at meter scale resolution only on the active part of the front to perform speed over 100Ha per second on a single CPU). It is inteded to serve both as a research platform and a tool for operational forecasting.
 
@@ -109,7 +109,7 @@ Each of these commands (such as `goTo[t=42]`, `include[state.ff]`, `startFire[lo
 
 ![Default web interface with data layers on the left pane, commands displayed as buttons and displaying an atmospheric coupled simulation of a wildfire in Portugal.\label{fig:gui}](gui.jpg)
 
-By utilizing pre-compiled datasets over extensive regions, this approach supports continent-wide operational forecasting services. It has been deployed to identify optimal escape routes [@kamilaris2023], integrated into the French National WildFire Decision Support System [OPEN DFCI](https://opendfci.fr/), showcased on the [FireCaster demonstration platform](https://forefire.univ-corse.fr/), and also currently used in commercial simulation services [AriaFire Firecaster](https://firecaster.ariafire.com), [UmGraueMeio Pantera](https://www.umgrauemeio.com/) and [Ororatech FireSpread](https://ororatech.com/all-products/fire-spread).
+By utilizing pre-compiled datasets over extensive regions, this approach supports continent-wide operational forecasting services. It has been deployed to identify optimal escape routes [@kamilaris2023], integrated into the French National WildFire Decision Support System [OPEN DFCI](https://opendfci.fr/), showcased on the [FireCaster demonstration platform](https://forefire.univ-corse.fr/), and also currently used in commercial simulation services [AriaFire Firecaster](https://firecaster.ariafire.com), [umgrauemeio Pantera](https://www.umgrauemeio.com/en) and [Ororatech FireSpread](https://ororatech.com/all-products/fire-spread).
 
 
 ### Two-way coupling with the MesoNH atmospheric model
