### 2025-11-27 c6420d7
replace double parenthesis for "— "
