### 2025-11-27 faa64c0
[joss] fixed typos (#140)
* typos and spacing

* other typos

* replace double parenthesis for "— "
