### 2025-11-28 4d01da0
Enhance clarity and readability in paper.md
Refined language and improved clarity in the summary and statement of need sections of the paper.

These changes are related to the JOSS review at https://github.com/openjournals/joss-reviews/issues/8680#issuecomment-3588328185
