+++ paper/paper.md
@@ -85,37 +85,37 @@ bibliography: paper.bib
 ---
 
 # Summary
-Wildfire forecasting is both an active research area and an important need for decision support systems. **ForeFire** is a modular, high-performance, scriptable, discrete‑event‑driven simulation engine [@filippi2009] focusing computational effort on the active region of a fire front defined as a dynamic mesh (or multipolygons) of fire markers. It is designed to model the spread of wildfire perimeters over large landscapes at meter scale resolution in seconds, serving both as a research platform and a tool for operational forecasting. The core C++ library has Fortran and Python bindings and is accompanied by a lightweight scriptable interpreter (a custom FF language), can load, save and export data in NetCDF, GeoJson, KML, PNG and JPG, and includes a local HTTP service with customizable graphical user interface. ForeFire can also account for fire-atmosphere interaction by two-way coupling with the MesoNH [@lac2018] atmospheric model [@filippi2013].
+Wildfire forecasting is both an active research area and an important need for decision support systems. **ForeFire** is a modular, high-performance, scriptable, discrete‑event‑driven simulation engine [@filippi2009] focusing computational effort on the active region of a fire front defined as a dynamic mesh (or multipolygons) of fire markers. It is designed to model the spread of wildfire perimeters over large landscapes at a meter scale resolution in seconds, serving both as a research platform and a tool for operational forecasting. The core C++ library has Fortran and Python bindings and is accompanied by a lightweight, scriptable interpreter (a custom FF language) that can load, save, and export data in NetCDF, GeoJSON, KML, PNG, and JPG, and includes a local HTTP service with a customizable graphical user interface. ForeFire can also account for fire-atmosphere interaction by two-way coupling with the MesoNH [@lac2018] atmospheric model [@filippi2013].
 
 # Statement of need
 
 Wildfire modeling tools have historically been split between **complex combustion research models** and **streamlined operational tools**, each with distinct limitations. 
-Computational combustion and fluid dynamics (CFD) based models (e.g., FIRETEC [@linn2005] or WFDS [@mell2007]) are highly computationally intensive and yet unable to provide large wildfire forecasts faster than real time. 
-Atmospheric coupled codes, such as WRF/SFire [@mandel2011] must be run within an atmospheric model and require a large amount of processing power and data.
-Operational wildfire simulators, such as widely used Farsite [@finney1998] (now Flammap [@Finney2023FlamMap]), or Canadian Prometheus [@garcia2008], are able to simulate fire fronts spanning tens of kilometers in a matter of seconds, but have definite built-in modeling assumptions and are distributed as compiled software with graphical interfaces with limited scriptability. Other open source libraries are ElmFire [@lautenberger2013] or Cell2Fire [@pais2021] that are tied to a single spread model and do not include scripting language, or deep learning based [@XIA2025106401]. 
+Computational combustion and fluid dynamics (CFD)- based models (e.g., FIRETEC [@linn2005] or WFDS [@mell2007]) are highly computationally intensive yet unable to provide large-scale wildfire forecasts faster than real time. 
+Atmospheric-coupled codes, such as WRF/SFire [@mandel2011], must be run within an atmospheric model and require substantial processing power and data.
[…]
