### 2025-11-28 4a21487
Improved clarity on PyTochFire sentence
