+++ paper/paper.md
@@ -92,7 +92,7 @@ Wildfire forecasting is both an active research area and an important need for d
 Wildfire modeling tools have historically been split between **complex combustion research models** and **streamlined operational tools**, each with distinct limitations. 
 Computational combustion and fluid dynamics (CFD)- based models (e.g., FIRETEC [@linn2005] or WFDS [@mell2007]) are highly computationally intensive yet unable to provide large-scale wildfire forecasts faster than real time. 
 Atmospheric-coupled codes, such as WRF/SFire [@mandel2011], must be run within an atmospheric model and require substantial processing power and data.
-Operational wildfire simulators, such as widely used Farsite [@finney1998] (now Flammap [@Finney2023FlamMap]), or Canadian Prometheus [@garcia2008], can simulate fire fronts spanning tens of kilometers in a matter of seconds, but have definite built-in modeling assumptions and are distributed as compiled software with graphical interfaces with limited scriptability. Other open-source libraries include ElmFire [@lautenberger2013] and Cell2Fire [@pais2021], which are tied to a single spread model and do not include a scripting language, or deep learning-based [@XIA2025106401]. 
+Operational wildfire simulators, such as widely used Farsite [@finney1998] (now Flammap [@Finney2023FlamMap]), or Canadian Prometheus [@garcia2008], can simulate fire fronts spanning tens of kilometers in a matter of seconds, but have definite built-in modeling assumptions and are distributed as compiled software with graphical interfaces with limited scriptability. Other open-source libraries include ElmFire [@lautenberger2013] and Cell2Fire [@pais2021], which are tied to a single spread model and do not include a scripting language. Deep-learning approaches also exist, such as PyTorchFire [@XIA2025106401].
 
 ForeFire was developed as a community tool to fill the gap between highly complex customizable models and more rigid operational tools: a **unified** wildfire simulator that is both **adaptable** (highly scriptable with multiple bindings) and **high-performing** (discrete‑event‑driven simulation with dynamic mesh allows to concentrate computation at meter scale resolution only on the active part of the front to perform speed over 100 Ha per second on a single CPU). It is intended to serve both as a research platform and a tool for operational forecasting.
 
