### 2025-11-29 #8680 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
