### 2025-12-03 #8680 @filippi
Many thanks to @philipcardiff , @kthyng , @xiazeyu, @fveigauvigo and @iperezx for the careful and constructive work throughout the review process.
The experience has been very positive, and it clearly shows how such an open and transparent review workflow can improve efficiency compared to more traditional systems.
I remain fully available for any next steps required on my side.
