### 2025-12-04 #8680 @kthyng
@filippi Can you update the metadata for your Zenodo archive so the version matches what you have here (`v2.4.2`)?
