### 2025-12-04 d7d87fb
paper updated refs
