+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{lautenberger2013,
-  title = {Wildland fire modeling with an Eulerian level set method and automated calibration},
+  title = {Wildland fire modeling with an {Eulerian} level set method and automated calibration},
   volume = {62},
   ISSN = {0379-7112},
   url = {https://github.com/lautenberger/elmfire/},
@@ -31,7 +31,7 @@ year = {1998}
   note         = {Available from U.S. Forest Service Fire Sciences Laboratory}
 }
 @article{XIA2025106401,
-title = {PyTorchFire: A GPU-accelerated wildfire simulator with Differentiable Cellular Automata},
+title = {{PyTorchFire}: A {GPU}-accelerated wildfire simulator with Differentiable Cellular Automata},
 journal = {Environmental Modelling & Software},
 volume = {188},
 pages = {106401},
@@ -43,7 +43,7 @@ author = {Zeyu Xia and Sibo Cheng},
 keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based software, Parallel computing techniques, GPU-acceleration},
 }
 @book{Scott2005,
-  title = {Standard fire behavior fuel models: a comprehensive set for use with Rothermel’s surface fire spread model},
+  title = {Standard fire behavior fuel models: a comprehensive set for use with {Rothermel’s} surface fire spread model},
   url = {http://dx.doi.org/10.2737/RMRS-GTR-153},
   DOI = {10.2737/rmrs-gtr-153},
   institution = {U.S. Department of Agriculture,  Forest Service,  Rocky Mountain Research Station},
@@ -205,7 +205,7 @@ volume = {156},
 year = {2009}
 }
 @book{andrews2018,
-  title = {The Rothermel surface fire spread model and associated developments: A comprehensive explanation},
+  title = {The {Rothermel} surface fire spread model and associated developments: A comprehensive explanation},
   url = {http://dx.doi.org/10.2737/RMRS-GTR-371},
   DOI = {10.2737/rmrs-gtr-371},
   institution = {U.S. Department of Agriculture,  Forest Service,  Rocky Mountain Research Station},
@@ -213,7 +213,7 @@ year = {2009}
   year = {2018}
 }
 @article{garcia2008,
-  title = {Smoothing and bootstrapping the PROMETHEUS fire growth model},
+  title = {Smoothing and bootstrapping the {PROMETHEUS} fire growth model},
   volume = {19},
   ISSN = {1099-095X},
   url = {http://dx.doi.org/10.1002/env.907},
@@ -227,7 +227,7 @@ year = {2009}
   pages = {836–848}
 }
 @article{pais2021,
-  title = {Cell2Fire: A Cell-Based Forest Fire Growth Model to Support Strategic Landscape Management Planning},
+  title = {{Cell2Fire}: A {Cell-Based} Forest Fire Growth Model to Support Strategic Landscape Management Planning},
   volume = {4},
   ISSN = {2624-893X},
   url = {http://dx.doi.org/10.3389/ffgc.2021.692706},
@@ -239,7 +239,7 @@ year = {2009}
   month = nov 
 }
 @article{mandel2011,
-  title = {Coupled atmosphere-wildland fire modeling with WRF  3.3 and SFIRE 2011},
+  title = {Coupled atmosphere-wildland fire modeling with {WRF} 3.3 and {SFIRE} 2011},
   volume = {4},
   ISSN = {1991-9603},
   url = {http://dx.doi.org/10.5194/gmd-4-591-2011},
@@ -254,7 +254,7 @@ year = {2009}
 }
 
 @article{campos2023, 
[…]
