### 2025-12-04 0609cf7
More checks,
