+++ paper/paper.bib
@@ -159,7 +159,7 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   number = {5},
   pages = {1929--1969},
   author = {Christine Lac and Jean-Pierre Chaboureau and Valéry Masson and Jean-Pierre Pinty and Pierre Tulet and Juan Escobar and Maud Leriche and Christelle Barthe and Benjamin Aouizerats and Clotilde Augros and Pierre Aumond and Franck Auguste and Peter Bechtold and Sarah Berthet and Soline Bielli and Frédéric Bosseur and Olivier Caumont and Jean-Martial Cohard and Jeanne Colin and Fleur Couvreux and Joan Cuxart and Gaëlle Delautier and Thibaut Dauhut and Véronique Ducrocq and Jean-Baptiste Filippi and Didier Gazen and Olivier Geoffroy and François Gheusi and Rachel Honnert and Jean-Philippe Lafore and Cindy Lebeaupin Brossier and Quentin Libois and Thibaut Lunet and Céline Mari and Tomislav Maric and Patrick Mascart and Maxime Mogé and Gilles Molinié and Olivier Nuissier and Florian Pantillon and Philippe Peyrillé and Julien Pergaud and Emilie Perraud and Joris Pianezze and Jean-Luc Redelsperger and Didier Ricard and Evelyne Richard and Sébastien Riette and Quentin Rodier and Robert Schoetter and Léo Seyfried and Joël Stein and Karsten Suhre and Marie Taufour and Odile Thouron and Sandra Turner and Antoine Verrelle and Beno{\^{\i}}t Vié and Florian Visentin and Vincent Vionnet and Philippe Wautelet},
-  title = {Overview of the Meso-{NH} model version 5.4 and its applications},
+  title = {Overview of the {Meso-NH} model version 5.4 and its applications},
   journal = {Geoscientific Model Development}
 }
 
@@ -195,7 +195,7 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   month = jan 
 }
 @article{balbi2009,
-author = {Balbi, J H and Morandini, F and Silvani, X and Filippi, J B and Rinieri, F},
+author = {Balbi, Jacques-Henri and Morandini, Frédéric and Silvani, Xavier and Filippi,  Jean-Baptiste and Rinieri, Frédéric},
 doi = {10.1016/j.combustflame.2009.07.010},
 journal = {Combustion and Flame},
 number = {12},
@@ -294,7 +294,7 @@ year = {2009}
 }
 @Article{filippi2021,
 	AUTHOR = {Filippi, Jean-Baptiste and Durand, Jonathan and Tulet, Pierre and Bielli, Soline},
-	TITLE = {Multiscale Modeling of Convection and Pollutant Transport Associated with Volcanic Eruption and Lava Flow: Application to the April 2007 Eruption of the {Piton de la Fournaise} ({Reunion Island})},
+	TITLE = {Multiscale Modeling of Convection and Pollutant Transport Associated with Volcanic Eruption and Lava Flow: Application to the {April} 2007 Eruption of the {Piton de la Fournaise} ({Reunion Island})},
 	JOURNAL = {Atmosphere},
 	VOLUME = {12},
 	YEAR = {2021},
@@ -306,7 +306,7 @@ year = {2009}
 	DOI = {10.3390/atmos12040507}
 }
 @article{filippi2013,
-    title = {Assessment of {FOREFIRE}/{MESONH} for wildland fire/atmosphere coupled simulation of the {FireFlux} experiment},
[…]
