### 2025-12-04 #8680 @filippi
Dear @kthyng
Thanks a lot for spotting those and for your editing work.
2 of us read again every word of the reference section, corrected the last ones and it seems OK (when capitalised it is matching original ref title)
We may still have left typos but we cannot see any more.
Thanks a lot
