+++ paper/paper.bib
@@ -8,7 +8,7 @@
   publisher = {Elsevier BV},
   author = {Lautenberger,  Chris},
   year = {2013},
-  month = nov,
+  month = Nov,
   pages = {289–298}
 }
 @techreport{finney1998,
@@ -74,7 +74,7 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   publisher = {Elsevier BV},
   author = {Kamilaris,  A. and Filippi,  J.B. and Padubidri,  C. and Koole,  R. and Karatsiolis,  S.},
   year = {2023},
-  month = apr,
+  month = Apr,
   pages = {103747}
 }
 
@@ -90,7 +90,7 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   publisher = {American Geophysical Union (AGU)},
   author = {Linn,  Rodman R. and Cunningham,  Philip},
   year = {2005},
-  month = jul 
+  month = Jul 
 }
 @article{allaire2022,
   title = {Simulation-based high-resolution fire danger mapping using deep learning},
@@ -103,7 +103,7 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   publisher = {CSIRO Publishing},
   author = {Allaire,  Frédéric and Filippi,  Jean-Baptiste and Mallet,  Vivien and Vaysse,  Florence},
   year = {2022},
-  month = apr,
+  month = Apr,
   pages = {379–394}
 }
 @article{allaire2021,
@@ -144,7 +144,7 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   publisher = {Copernicus GmbH},
   author = {Filippi,  Jean-Baptiste and Mallet,  Vivien and Nader,  Bahaa},
   year = {2014},
-  month = nov,
+  month = Nov,
   pages = {3077–3091}
 }
 
@@ -176,7 +176,7 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   publisher = {SAGE Publications},
   author = {Filippi,  Jean-Baptiste and Morandini,  Frédéric and Balbi,  Jacques Henri and Hill,  David RC},
   year = {2009},
-  month = aug,
+  month = Aug,
   pages = {629–646}
 }
 
@@ -192,7 +192,7 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   author = {Santoni,  Paul-Antoine and Filippi,  Jean-Baptiste and Balbi,  Jacques-Henri and Bosseur,  Frédéric},
   editor = {Morvan,  D},
   year = {2011},
-  month = jan 
+  month = Jan 
 }
 @article{balbi2009,
 author = {Balbi, Jacques-Henri and Morandini, Frédéric and Silvani, Xavier and Filippi,  Jean-Baptiste and Rinieri, Frédéric},
@@ -223,7 +223,7 @@ year = {2009}
   publisher = {Wiley},
   author = {Garcia,  Tanya and Braun,  John and Bryce,  Robert and Tymstra,  Cordy},
   year = {2008},
-  month = feb,
+  month = Feb,
   pages = {836–848}
 }
 @article{pais2021,
@@ -249,7 +249,7 @@ year = {2009}
   publisher = {Copernicus GmbH},
   author = {Mandel,  J. and Beezley,  J. D. and Kochanski,  A. K.},
   year = {2011},
-  month = jul,
+  month = Jul,
   pages = {591–610}
 }
 
@@ -263,7 +263,7 @@ year = {2009}
   publisher = {Elsevier BV},
[…]
