+++ paper/paper.bib
@@ -60,8 +60,7 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   journal = {International Journal of Wildland Fire},
   publisher = {CSIRO Publishing},
   author = {Mell,  William and Jenkins,  Mary Ann and Gould,  Jim and Cheney,  Phil},
-  year = {2007},
-  pages = {1}
+  year = {2007}
 }
 
 @article{kamilaris2023,
@@ -85,12 +84,11 @@ keywords = {Wildfire simulation, Differentiable Cellular Automata, PyTorch-based
   ISSN = {0148-0227},
   url = {http://dx.doi.org/10.1029/2004JD005597},
   DOI = {10.1029/2004jd005597},
-  number = {D13},
+  number = {D13107},
   journal = {Journal of Geophysical Research: Atmospheres},
   publisher = {American Geophysical Union (AGU)},
   author = {Linn,  Rodman R. and Cunningham,  Philip},
-  year = {2005},
-  month = Jul 
+  year = {2005}
 }
 @article{allaire2022,
   title = {Simulation-based high-resolution fire danger mapping using deep learning},
@@ -322,7 +320,7 @@ year = {2009}
 @article{strada2012,
   title = {Wildfire and the atmosphere: Modelling the chemical and dynamic interactions at the regional scale},
   volume = {51},
-  ISSN = {1352-2310},
+  ISSN = {1352-2310},grassland
   url = {http://dx.doi.org/10.1016/j.atmosenv.2012.01.023},
   DOI = {10.1016/j.atmosenv.2012.01.023},
   journal = {Atmospheric Environment},
