+++ paper/paper.bib
@@ -320,7 +320,7 @@ year = {2009}
 @article{strada2012,
   title = {Wildfire and the atmosphere: Modelling the chemical and dynamic interactions at the regional scale},
   volume = {51},
-  ISSN = {1352-2310},grassland
+  ISSN = {1352-2310},
   url = {http://dx.doi.org/10.1016/j.atmosenv.2012.01.023},
   DOI = {10.1016/j.atmosenv.2012.01.023},
   journal = {Atmospheric Environment},
