### 2025-12-05 #8680 @filippi
dear @kthyng , we made a last check on bib and md file, We now hope most if not all typos are gone.
Thanks a lot again for your editing work
